Split core.config.js out of core.controller.js
